refactor(analytics): split query module, TPS excludes recovered attempts, renderer column dedupe
- analytics-queries.ts (2,018 lines) split into aggregates + analytics-detail-
  queries.ts (drill-downs) + analytics-query-shared.ts (filters, Decimal/cost
  streaming, latency, name resolution); all previous exports re-exported, so
  worker/runner/IPC/tests needed no import changes (closes review #25)
- sessionNameResolver(ctx, db) replaces the 6x ctx-default resolver expression
  (closes review #31)
- TPS ratio now counts only succeeded attempts: crash-recovered rows carry a
  sentinel completed_at (recovery/disconnect time) that diluted the
  token-weighted rate; TTFT still counts every stamped attempt. Fixes #159
  (option 1). Latency SELECTs unified behind latencyRowColumns/Gating helpers
  and now carry outcome
- renderer: shared Column<T>, maxCostAmount, dateSortValue replace ~9 cost
  sort expressions and three inconsistent date-sort conventions across tabs
  (closes review #26)
- tests: interrupted-attempt TPS exclusion (null alone, excluded when mixed),
  maxCostAmount/dateSortValue units
