+++ electron/tests/unit/analytics-presentation.test.ts
@@ -9,6 +9,8 @@ import {
   formatDate,
   formatTps,
   formatTtft,
+  dateSortValue,
+  maxCostAmount,
   netInputTokens,
   netOutputTokens,
   tokenStackTooltipRows,
@@ -75,5 +77,47 @@ describe('stacked-token netting helpers', () => {
     expect(formatTtft(0)).toBe('—');
     expect(formatTtft(450)).toBe('450ms');
     expect(formatTtft(2345)).toBe('2.3s');
+  });
+});
+
+describe('analytics table sort helpers', () => {
+  it('maxCostAmount returns the largest cost across currencies, empty → 0', () => {
+    expect(maxCostAmount([])).toBe(0);
+    expect(maxCostAmount([{ currency: 'USD', amount: '1.25' }])).toBe(1.25);
+    expect(maxCostAmount([
+      { currency: 'EUR', amount: '1.5' },
+      { currency: 'USD', amount: '3.75' },
+      { currency: 'JPY', amount: '2' },
+    ])).toBe(3.75);
+  });
+
+  it('maxCostAmount clamps negatives so they never lead a descending sort', () => {
+    expect(maxCostAmount([{ currency: 'USD', amount: '-2' }])).toBe(0);
+    expect(maxCostAmount([
+      { currency: 'USD', amount: '-2' },
+      { currency: 'EUR', amount: '0.5' },
+    ])).toBe(0.5);
+  });
+
+  it('dateSortValue maps null and unparseable strings to 0 and valid ISO to positive ms', () => {
+    expect(dateSortValue(null)).toBe(0);
+    expect(dateSortValue('not-a-date')).toBe(0);
+    expect(dateSortValue('2026-01-01T00:30:00.000Z')).toBe(Date.parse('2026-01-01T00:30:00.000Z'));
+    expect(dateSortValue('2026-01-01T00:30:00.000Z')).toBeGreaterThan(0);
+  });
+
+  it('dateSortValue orders ISO-8601 stamps chronologically, matching lexicographic order', () => {
+    const stamps = [
+      '2026-01-02T00:00:00.000Z',
+      '2025-12-31T23:59:59.999Z',
+      '2026-06-15T10:20:30.400Z',
+      '2026-01-02T00:00:00.100Z',
+    ];
+    const byLexicographic = [...stamps].sort();
+    const byDateSortValue = [...stamps].sort((a, b) => dateSortValue(a) - dateSortValue(b));
+    expect(byDateSortValue).toEqual(byLexicographic);
+
+    // null sorts as 0 — below any post-epoch timestamp.
+    expect(dateSortValue(null)).toBeLessThan(dateSortValue(stamps[1]));
   });
 });
